# models/modelcrud/images.delete.go
// Code generated by SQLBoiler 4.6.0 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
// This file is meant to be re-generated in place and/or deleted at any time.

package modelcrud

import (
	"net/http"

	"github.com/calebhiebert/go-vue-template/api"
	"github.com/calebhiebert/go-vue-template/models"
	"github.com/gin-gonic/gin"
	"github.com/volatiletech/null/v8"
	"github.com/volatiletech/sqlboiler/v4/boil"
	"github.com/volatiletech/sqlboiler/v4/queries/qm"
)

// DeleteImageByID godoc
// @Summary Soft deletes a single Image entity based on their id
// @Produce json
// @Success 200 {object} APIImage
// @Param id path string true "Image id"
// @Param hardDelete query string false "Hard delete image"
// @Router /crud/images/:id [delete]
func (*GeneratedCrudController) DeleteImageByID(c *gin.Context) {
	id := c.Param("id")

	if id == "" {
		api.NewAPIError("invalid-id", http.StatusBadRequest, "The provided id was invalid").Respond(c)
		return
	}

	hardDelete := c.Query("hardDelete") == "true"

	existingImage, err := models.Images(qm.Where("id = ?", id)).OneG(c.Request.Context())
	if err != nil {
		api.APIErrorFromErr(err).Respond(c)
		return
	}

	_, err = existingImage.DeleteG(c.Request.Context(), hardDelete)
	if err != nil {
		api.APIErrorFromErr(err).Respond(c)
		return
	}

	c.JSON(http.StatusOK, existingImage)
}

// BulkDeleteImagesByIDs godoc
// @Summary Soft deletes a range of images by their ids
// @Produce json
// @Success 200 {object} DeletedCount
// @Param req body IDList true "List of ids to delete"
// @Param hardDelete query string false "Hard delete image"
// @Router /crud/images [delete]
func (*GeneratedCrudController) BulkDeleteImagesByIDs(c *gin.Context) {

	var ids IDList

	err := c.BindJSON(&ids)
	if err != nil {
		api.APIErrorFromErr(err).Respond(c)
		return
	}

	hardDelete := c.Query("hardDelete") == "true"

	var idInterface []interface{}

	for _, id := range ids.IDs {
		idInterface = append(idInterface, id)
	}

	deleted, err := models.Images(qm.WhereIn("id IN ?", idInterface...)).DeleteAllG(c.Request.Context(), hardDelete)
	if err != nil {
		api.APIErrorFromErr(err).Respond(c)
		return
	}

	c.JSON(http.StatusOK, DeletedCount{DeletedCount: int(deleted)})
}

// UnDeleteImageByID godoc
// @Summary Undeletes a image by id
// @Produce json
// @Success 200 {object} APIImage
// @Param id path string true "Image id"
// @Router /crud/images/:id/unDelete [post]
func (*GeneratedCrudController) UnDeleteImageByID(c *gin.Context) {
	id := c.Param("id")

	if id == "" {
		api.NewAPIError("invalid-id", http.StatusBadRequest, "The provided id was invalid").Respond(c)
		return
	}

	deletedImage, err := models.Images(qm.Where("id = ?", id), qm.WithDeleted()).OneG(c.Request.Context())
	if err != nil {
		api.APIErrorFromErr(err).Respond(c)
		return
	}

	deletedImage.DeletedAt = null.Time{
		Valid: false,
	}

	_, err = deletedImage.UpdateG(c.Request.Context(), boil.Whitelist("deleted_at"))
	if err != nil {
		api.APIErrorFromErr(err).Respond(c)
		return
	}

	c.JSON(http.StatusOK, deletedImage)
}